Several other inner classes moved to search subpkg
